Rename stale() to retire(); drop the successor promise
No clear use case yet for the successor value, so the callback takes no
argument and retire() resolves with void. Still fires on retirement —
re-render or unmount.
