+++ CHANGELOG.md
@@ -26,15 +26,13 @@
   expression, or the template edge is still stripped as layout.
 
 ### New API
-- **`Context.stale(callback?)`** (#324)
+- **`Context.retire(callback?)`** (#324)
   Registers a callback that fires when the current render is *retired* —
   superseded by a re-render, or removed when the component unmounts — so a
   single callback tears down the retired render's in-flight async work (e.g.
-  aborting a fetch via `AbortController`) in both cases. The callback receives a
-  promise for the successor render's result: it resolves with the next rendered
-  value on re-render, or `undefined` on unmount. Called with no argument,
-  `stale()` returns that promise directly. Re-register it on each render; it
-  never itself triggers a re-render.
+  aborting a fetch via `AbortController`) in both cases. Called with no argument,
+  `retire()` returns a promise that resolves when the render is retired.
+  Re-register it on each render; it never itself triggers a re-render.
 
 ### Changed
 - **`window.Crank` (the UMD/CDN browser global) now ships authoring templates and a default renderer.**
