+++ src/crank.ts
@@ -2287,52 +2287,18 @@ const provisionMaps = new WeakMap<ContextState, Map<unknown, unknown>>();
 
 const scheduleMap = new WeakMap<ContextState, Set<Function>>();
 
-const staleMap = new WeakMap<ContextState, Set<Function>>();
-
-// Pending stale() successor resolvers, drained at the next commit (with the
-// successor render's result) or on unmount (with undefined).
-const successorResolverMap = new WeakMap<
-	ContextState,
-	Array<(result: unknown) => void>
->();
+const retireMap = new WeakMap<ContextState, Set<Function>>();
 
 const cleanupMap = new WeakMap<ContextState, Set<Function>>();
 
 // A render is "retired" when it is superseded by a re-render or removed by an
-// unmount. Fire its stale() callbacks, handing each a promise for the successor
-// render's result — or undefined when the component unmounts (no successor).
-function retire(ctx: ContextState, isUnmount: boolean): void {
-	const callbacks = staleMap.get(ctx);
-	if (!callbacks) {
-		return;
-	}
-
-	staleMap.delete(ctx);
-	for (const callback of callbacks) {
-		if (isUnmount) {
-			callback(Promise.resolve(undefined));
-		} else {
-			let resolve!: (result: unknown) => void;
-			const successor = new Promise<unknown>((r) => (resolve = r));
-			let resolvers = successorResolverMap.get(ctx);
-			if (!resolvers) {
-				resolvers = [];
-				successorResolverMap.set(ctx, resolvers);
-			}
-
-			resolvers.push(resolve);
-			callback(successor);
-		}
-	}
-}
-
-// Resolve any pending stale() successor promises with the given result.
-function resolveSuccessors(ctx: ContextState, result: unknown): void {
-	const resolvers = successorResolverMap.get(ctx);
-	if (resolvers) {
-		successorResolverMap.delete(ctx);
-		for (const resolve of resolvers) {
-			resolve(result);
+// unmount. Fire its retire() callbacks.
+function retire(ctx: ContextState): void {
+	const callbacks = retireMap.get(ctx);
+	if (callbacks) {
+		retireMap.delete(ctx);
+		for (const callback of callbacks) {
+			callback();
 		}
 	}
 }
@@ -2618,8 +2584,8 @@ export class Context<
 		const schedulePromises: Array<PromiseLike<unknown>> = [];
 		try {
 			setFlag(ctx.ret, IsRefreshing);
-			// An explicit re-render retires the current render: fire stale() first.
-			retire(ctx, false);
+			// An explicit re-render retires the current render: fire retire() first.
+			retire(ctx);
 			diff = enqueueComponent(ctx);
 			if (isPromiseLike(diff)) {
 				return diff
@@ -2726,35 +2692,29 @@ export class Context<
 	/**
 	 * Registers a callback which fires when the current render is *retired* —
 	 * superseded by a re-render, or removed when the component unmounts. Useful
-	 * for tearing down the retired render's in-flight async work before it goes
-	 * stale, e.g. aborting a fetch via `AbortController`, in both the re-render
-	 * and unmount cases with a single callback.
+	 * for tearing down the retired render's in-flight async work, e.g. aborting a
+	 * fetch via `AbortController`, in both the re-render and unmount cases with a
+	 * single callback.
 	 *
 	 * The callback fires once per registration; re-register it on each render
 	 * (typically inside the `for...of`/`for await...of` loop). It never itself
 	 * triggers a re-render.
 	 *
-	 * The callback is passed a promise for the *successor* render's result: it
-	 * resolves with the next rendered value when the component re-renders, or
-	 * with `undefined` when the component unmounts (no successor). Called with no
-	 * argument, `stale()` returns that same successor promise directly.
+	 * Called with no argument, returns a promise that resolves when the render is
+	 * retired.
 	 */
-	stale(): Promise<TResult | undefined>;
-	stale(callback: (successor: Promise<TResult | undefined>) => unknown): void;
-	stale(
-		callback?: (successor: Promise<TResult | undefined>) => unknown,
-	): Promise<TResult | undefined> | void {
+	retire(): Promise<void>;
+	retire(callback: () => unknown): void;
+	retire(callback?: () => unknown): Promise<void> | void {
 		if (!callback) {
-			return new Promise<TResult | undefined>((resolve) =>
-				this.stale((successor) => resolve(successor)),
-			);
+			return new Promise<void>((resolve) => this.retire(() => resolve()));
 		}
 
 		const ctx = this[_ContextState];
-		let callbacks = staleMap.get(ctx);
+		let callbacks = retireMap.get(ctx);
 		if (!callbacks) {
 			callbacks = new Set<Function>();
-			staleMap.set(ctx, callbacks);
+			retireMap.set(ctx, callbacks);
 		}
 
 		callbacks.add(callback);
@@ -2888,8 +2848,8 @@ function diffComponent<TNode, TScope, TRoot extends TNode | undefined, TResult>(
 		}
 
 		// A mounted component is about to re-render (e.g. from a prop change or a
-		// parent re-render), retiring the current render: fire its stale() callbacks.
-		retire(ctx, false);
+		// parent re-render), retiring the current render: fire its retire() callbacks.
+		retire(ctx);
 	} else {
 		ctx = ret.ctx = new ContextState(adapter, root, host, parent, scope, ret);
 	}
@@ -3552,10 +3512,7 @@ function commitComponent<TNode>(
 	// We always use getValue() instead of the unwrapping values because there
 	// are various ways in which the values could have been updated, especially
 	// if schedule callbacks call refresh() or async mounting is happening.
-	const committed = getValue(ctx.ret, true);
-	// This render is the successor of a retired one: resolve its stale() promises.
-	resolveSuccessors(ctx, ctx.adapter.read(committed));
-	return committed;
+	return getValue(ctx.ret, true);
 }
 
 /**
@@ -3654,10 +3611,8 @@ async function unmountComponent(
 		return;
 	}
 
-	// Unmount retires the current render with no successor: fire stale()
-	// callbacks (undefined successor) and settle any dangling successor promises.
-	retire(ctx, true);
-	resolveSuccessors(ctx, undefined);
+	// Unmount retires the current render: fire retire() callbacks.
+	retire(ctx);
 
 	let cleanupPromises: Array<PromiseLike<unknown>> | undefined;
 	// TODO: think about errror handling for callbacks
